Well we did alot so far
I decided that the best way to handle the index file was to read it in
on the beginning of a transaction and then invalidiate it and resave the
whole thing at the end of the transaction. It isn't the best way but
it bets seeking through an ASCII file which is not worth the time or
resources.